Refactoring TextField
