+++ superformula-frontend/src/components/TextField/TextField.stories.tsx
@@ -3,8 +3,7 @@ import { Story } from '@storybook/react'
 import { Meta } from '@storybook/react/types-6-0'
 
 import styles from '../../styles/themes.module.scss'
-import { TextFieldProps } from './TextField.types'
-import TextField from './TextField'
+import TextField, { TextFieldProps } from './TextField'
 
 export default {
     title: 'Components/TextField',
@@ -21,4 +20,4 @@ export default {
 const Template: Story<TextFieldProps> = (args: TextFieldProps) => <TextField {...args} />
 
 export const Main = Template.bind({})
-Main.args = { label: 'label', inputId: 'inputId' }
+Main.args = { label: 'Label', inputId: 'inputId', placeholder: 'Placeholder' }
