+++ superformula-frontend/src/components/TextField/TextField.tsx
@@ -1,17 +1,22 @@
 import React, { ReactElement } from 'react'
-import { TextFieldProps } from './TextField.types'
 import styles from './TextField.module.scss'
-import Input from '../Input/Input'
+import Input, { InputProps } from '../Input/Input'
+
+type TextFieldProps = {
+    label: string
+    inputId: string
+} & InputProps
 
 function TextField({ label, inputId, ...props }: TextFieldProps): ReactElement {
     return (
-        <div className={styles.textField}>
+        <>
             <label className={styles.label} htmlFor={inputId}>
                 {label}
             </label>
             <Input className={styles.input} id={inputId} {...props} />
-        </div>
+        </>
     )
 }
 
 export default TextField
+export type { TextFieldProps }
