--- superformula-frontend/src/components/TextField/TextField.types.tsx
@@ -1,8 +1,0 @@
-import { InputProps } from 'components/Input'
-
-type TextFieldProps = {
-    label: string
-    inputId: string
-} & InputProps
-
-export type { TextFieldProps }
